requirements.txt
altair==4.1.0
pandas==1.2.4
Pillow==8.2.0
Pillow==9.5.0
streamlit==1.22.0
